requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
python-telegram-bot
pyyaml
python-dotenv